bruh jsut did erythin

+++ src/components/Posts.vue
@@ -12,24 +12,30 @@
         </div>
       </router-link>
 
-      <i v-if="account.id == post.creator.id" class="mdi yo fs-6 mdi-delete btn selectable p-0"
+      <i v-if="account.id == post.creator.id" title="Delete" class="mdi yo fs-6 mdi-delete btn selectable p-0"
         @click="deletePost(post.id)"></i>
     </div>
     <div class="card-body d-flex flex-column">
       {{post.body}}
-
     </div>
     <div class="card-footer mb-2 rounded" v-if="post.imgUrl">
       <img class="img-fluid mb-1 rounded" :src="post.imgUrl" alt="Broken img or sum">
     </div>
-    <span v-if="checkLike == true" class="px-3 pb-3">
-      <i class="mdi selectable rounded fs-2 mdi-heart" @click="changeLike(post.id)"></i>
-      {{post.likeAmount}}
-    </span>
-    <span v-else="" class="px-3 pb-3">
-      <i class="mdi selectable rounded fs-2 mdi-heart-outline" @click="changeLike(post.id)"></i>
-      {{post.likeAmount}}
-    </span>
+    <div class="d-flex justify-content-between align-items-center">
+      <span v-if="checkLike == true" class="px-3 pb-3">
+        <i title="Unlike" class="mdi selectable rounded fs-2 mdi-heart" @click="changeLike(post.id)"></i>
+        {{post.likeAmount}}
+      </span>
+      <span v-else="" class="px-3 pb-3">
+        <i title="Like" class="mdi selectable rounded fs-2 mdi-heart-outline" @click="changeLike(post.id)"></i>
+        {{post.likeAmount}}
+      </span>
+
+      <span>
+        Posted: {{this.time}}
+
+      </span>
+    </div>
   </div>
 </template>
 
@@ -37,6 +43,7 @@
 <script>
 // import { computed } from '@vue/reactivity';
 import { computed } from '@vue/reactivity';
+import { onMounted } from 'vue';
 import { AppState } from '../AppState.js';
 // import { Account } from '../models/Account.js';
 import { Post } from '../models/Post.js';
@@ -48,9 +55,18 @@ export default {
     post: { type: Post, required: true },
   },
   setup(props) {
-
+    function test() {
+      const tes = new Date(props.post.createdAt)
+      // console.log(tes.getDay());
+      // console.log(Math.floor(tes.getTime() / 1000.0))
+    }
+    onMounted(() => {
+      test()
+    })
     return {
-
+      // TODO why is this so annoying
+      time: (new Date(props.post.createdAt)),
+      // time1: time.getDay,
       account: computed(() => AppState.account),
       checkLike: computed(() => {
         let yo = false
@@ -62,7 +78,11 @@ export default {
         return yo
       }),
       async deletePost(id) {
-        console.log(id);
+        // console.log(id);
+        const yes = await Pop.confirm('Delete This Post?')
+        if (!yes) {
+          return
+        }
         try {
           await postService.deletePost(id)
         } catch (error) {
